Frontend + XML API Functions
Made a list of functions that I need working for 100% functionality of the app.

I have been going through that list and creating functions one by one.

Figured out and tested about 8 functions in this commit.

For some more context: The XML API that I am calling is not well documented. The documentation that does exist is 8 years old. Moreover, incorrect usage crashes the software due to memory exception errors.